+++ session/src/init/templates/routes/login_basic.tsx
@@ -1,3 +1,4 @@
+import { Button } from "../components/Button.tsx";
 import { define } from "../utils.ts";
 
 export const handler = define.handlers({
@@ -9,7 +10,7 @@ export const handler = define.handlers({
       // Logic to verify credentials goes here
 
       // Log user in
-      await ctx.state.login(username, { username });
+      await ctx.state.login(username);
       return ctx.redirect("/");
     }
 
@@ -46,9 +47,9 @@ export default define.page<typeof handler>((ctx) => {
             required
           />
         </div>
-        <button type="submit" class="bg-blue-500 text-white p-2 rounded">
+        <Button type="submit" class="bg-blue-500 text-white p-2 rounded w-full">
           Login
-        </button>
+        </Button>
       </form>
 
       <p class="mt-4 text-center">