+++ session/src/init/templates/routes/register_prod.tsx
@@ -1,3 +1,4 @@
+import { Button } from "../components/Button.tsx";
 import { define } from "../utils.ts";
 /* import { hash } from "@felix/argon2"; */
 
@@ -34,7 +35,7 @@ export const handler = define.handlers({
     */
 
     // 3. Log them in
-    await ctx.state.login(username, { username });
+    await ctx.state.login(username);
 
     return ctx.redirect("/");
   },
@@ -76,9 +77,12 @@ export default define.page<typeof handler>((ctx) => {
             required
           />
         </div>
-        <button type="submit" class="bg-green-500 text-white p-2 rounded">
+        <Button
+          type="submit"
+          class="bg-green-500 text-white p-2 rounded w-full"
+        >
           Register
-        </button>
+        </Button>
       </form>
 
       <p class="mt-4 text-center">